Extract test functions to a shared_functions file.  Please enter the commit message for your changes. Lines starting
